+++ src/app/_layout.tsx
@@ -97,6 +97,10 @@ export default function RootLayout() {
                         options={{ headerShown: false }}
                     />
                     <Stack.Screen
+                        name="pages/user/profile/index"
+                        options={{ headerShown: false }}
+                    />
+                    <Stack.Screen
                         name="auth/welcome"
                         options={{ headerShown: false }}
                     />